Create Word_counter
A word counter is used in used at the back end in Microsoft word and a lot of other software's. I went a step further to count how many times each word was repeated. This is used in Sentiment Analysis to computationally identify and categorize opinions expressed in a piece of text, especially in order to determine whether the writer's attitude towards a particular topic, product, etc. is positive, negative, or neutral. It is also used to rate products after running Machine Learning algorithms to rate products from customer reviews.